refactor(web): move password input to design system
