+++ web/src/pages/project/[projectId]/settings/integrations/posthog.tsx
@@ -13,7 +13,7 @@ import {
   FormMessage,
 } from "@/src/components/ui/form";
 import { Input } from "@/src/components/ui/input";
-import { PasswordInput } from "@/src/components/ui/password-input";
+import { PasswordInput } from "@/src/components/design-system/PasswordInput/PasswordInput";
 import { Switch } from "@/src/components/design-system/Switch/Switch";
 import {
   Select,
